Machine Learning Lab/requirements.txt
numpy==1.24.3
Pillow==9.5.0
Pillow==10.2.0
streamlit==1.26.0
